// apps/api/src/services/pim/channel-specs/etsy-listing-schema.ts
export const etsyListingSchema = {
  source: "https://www.etsy.com/openapi/generated/oas/3.0.0.json",
  verifiedAt: "2026-09-10",
  listing: { type: "object", properties: {
// ...
    "listing_type": {"type":"string","enum":["physical","download","both"]},
// ...
    "style": {"type":"array","items":{"type":"string"}},
// ...
  } },
  create: { type: "object", required: ["quantity","title","description","price","who_made","when_made","taxonomy_id"], properties: {
// ...
    "styles": {"type":"array","nullable":true,"items":{"type":"string"}},
// ...
    "type": {"type":"string","enum":["physical","download","both"]}
  } },
// ...
} as const